Deleted LinkedList
